requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-bootstrap-form==3.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-nose==1.4.6
djangorestframework==3.9.0
idna==2.8
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
nose==1.3.7
plumbum==1.6.7
pytz==2018.7
requests==2.21.0
rpyc==4.0.2
urllib3==1.24.1
uWSGI==2.0.17.1